[ST40] Add support for boot-from-NAND on the H225-HDK (B2042) board
This commit adds support for boot-from-NAND on STMicroelectronics'
STiH225-HDK (B2042) board, which uses STMicroelectronics' STiH225
SoC device.

This commit is a faithful recapitulation of the original patch
(attachment #15301) in Bugzilla #16539. For details, please see:

	https://bugzilla.stlinux.com/show_bug.cgi?id=16539
	https://bugzilla.stlinux.com/attachment.cgi?id=15301

This commit has not been tested, other than that stated in the
aforementioned bugzilla history, save it building.

Thanks to Jon, for the original patch.

Signed-off-by: Jon Frosdick <jon.frosdick@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>