[ST40] Confirm boot-from-NAND+NOR for STiH225-HDK (B2042)
This commit simply adds a useful comment to include/configs/hdkh225.h, and
does not logically change the code. The comment is to include specific
details explaining how to configure the switches on the STiH225-HDK
(B2042) board for both boot-from-NOR, and boot-from-NAND environments.
The comment also explains that a board modification is required if
boot-from-NAND is required (specifically MODE[11]).

The comment is, in essence, a more comprehensive summary of the details
provided in https://bugzilla.stlinux.com/show_bug.cgi?id=16539

In addition, some testing has been done to confirm that boot-from-NOR,
and boot-from-NAND does work correctly with this commit. Both NAND and
NOR both being bootable, and modifiable irrespective of the current
boot device.

In addition, networking with the on-board IC+ IP1001 PHY (on J24,
using GMAC #0) was confirmed as working satisfactorily. USB has *not*
yet been tested by the submitter, neither has the other RJ-45 port
(on EJ1, using GMAC #1) which uses the in-package PHY.

All testing was performed on a single board "STiH225-HDK10ST E004".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>