[ST40] Add support for boot-from-SPI on B2042 (STiH225-HDK)
This commit adds support for SPI (including boot-from-SPI) on the B2042
(STiH225-HDK) board, which uses STMicroelectronics' STiH225 SoC device.

This particular SoC, does *not* include the new SPI "FSM" H/W controller,
hence SPI support has been added only for use with the S/W "bit-banging"
driver, and for the H/W SSC controller driver - both of which can be
used to read+write the SPI Serial Flash device.

The configuration file "include/configs/hdkh225.h" has been updated
to expand the comment describing the jumper+switch configurations,
specifically to boot from SPI (in addition to NAND and NOR).

To boot-from-SPI, the following macro should be defined:

	#define CFG_BOOT_FROM_SPI

By default, this macro is *not* defined - the default boot-device is
currently set to boot from NOR.

To select between the "bit-banging" and the SSC drivers, one of the
following macros should be defined:

	#define CONFIG_STM_SSC_SPI	/* use SSC */
	#define CONFIG_SOFT_SPI		/* use bit-banging */

The default is to use the H/W SSC, which is much faster!

As well as reading and writing to the SPI Serial Flash device, the board
was booted directly from SPI (after changing the switches+jumpers). Both
the "bit-banging" and the SSC drivers were tested, and no issues were
noted. The SSC was only tested as 5MHz, it *may* be possible to speed
this clock up slightly.

All testing was performed on a single board "STiH225-HDK10ST E004".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>