[ST40,FIX] Ethernet was broken after USB started on STx7141 boards
This commit fixes a bug in the USB initialization code, specifically for
STMicroelectronics' STx7141 SoC, which accidentally disabled the on-die
GMACs, and hence disabled ethernet, when "usb start" was performed.

Essentially, the code was meant to set the bottom nibble of SYS_CFG07
to 0xA (bits[3:2]=0x2, and bits[1:0]=0x2), but it incorrectly set the
whole of the 32-bit register to 0x0000000A, disabling the GMACs as a
by-product, The fix is simply to set the bottom nibble to 0xA, whilst
preserving the top 28-bits of the SYS_CFG07 register.

Testing was simply running the "nfs" command and the "usb start" command
(and friends) in various combinations, ensuring that USB and Ethernet
both continued to work, as expected.

All testing was performed on a single board "STiH225-HDK10ST E004".

For further details, please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=10991
	https://bugzilla.stlinux.com/show_bug.cgi?id=18844

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>