[ST40] Add initial support for the B2064 Board (STxH239)
This commit adds initial support for the B2064 board from
STMicroelectronics, which is also known as the STiH239-HDK board, which
has a STiH239 SoC device (a derivative of the STiH205/STiH207).

This board is very similar to previously supported boards (the B2057, and
the B2067).  It is so similar that it was decided to reuse the existing
board/st/b2057 directory and the files therein for it, and not create a
separate directory (board/st/b2064), as one would normally do to support
a new board.

Only basic testing of this support has been done, and only on a pair of
B2064 boards:
	B2064A board (STiH239-HDKA003 E009)
	B2064B board (STiH239-HDKB200 E075)

Testing has included support for: Networking (with in-package IC+
IP101G PHY); USB (OHCI only); SPI serial FLASH; and multi-bit ECC
NAND (MT29F4G08ABADAH4) using the BCH controller in 18-bit ECC mode.
In addition, both boot-from-SPI, and boot-from-NAND has been tested,
and the board was able to boot from both FLASH types with no problems.

NAND was only tested with 18-bit ECC on the BCH controller, as the fitted
NAND device is not capable of supporting the ECC needs for 30-bit ECC,
and using a weaker ECC is not recommended!

USB was tested successfully using only the rear port (USB #0, JD1) on the
B2064A board. It is believed that there is a H/W issue on the specific
board under test which resulted in the front port (USB #1, JD3) not
working correctly. However, on the B2064B board, both the front and the
rear USB ports were demonstrably working as expected (non concurrently).

The same binaries may be used on either the 'A' or the 'B' revision.
That is, there is no obvious software differences needed by U-Boot to
run on either board revision.

Note: Please ensure you read the comments near the beginning of the
"include/configs/b2064.h" file for details about setting up the mode pins
correctly. The ones in this file have been empirically verified for the
specific NAND device on the board (MT29F4G08ABADAH4), which is a 512 MiB
NAND, with a page size of 2048+64 bytes and a sector size of 128 KiB,
requiring 5 address cycles. If you are using a different NAND device,
then you may need to modify the mode-pins to boot from it - check the
SoC's data-sheet for details.

Note: It is recommended that the romgen utility is re-run and the file
"board/st/b2057/b2064-stxh205.romgen" is replaced, when building
an image to boot from FLASH. Please read the comments in the file
"board/st/b2057/init-b2064.S" for details on how to run romgen.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>