[ST40] Updated TargetPack (poke-table) to "v09" for B2067
This commit replaces the previous romgen-generated "poke-table" file
"b2067-stxh205.romgen", with a new one, based on the latest (as of the
time of this commit) version of the master "TargetPack", which is "v09".

Previously, the "poke-table" was derived from version "v06", which
proved to demonstrate an element of instability with the DDR RAM. This
commit simply replaces the old poke-table for the B2067 board for a more
stable one.

Note: is is still always recommended that (especially for production
systems) users heed the standard advise of: re-running "romgen"
themselves with the latest version of the TargetPacks are available,
at that juncture.  Recall, in general, STMicroelectronics does *not*
routinely update the "*.romgen" files supplied in the U-Boot source tree.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>