[ST40] Fixed the "reset" issue on the STxH205/207 family
This commit fixes a inherent bug pertaining to the (previously)
broken "reset" command that affected all boards with a SoC from
STMicroelectronics' STxH205/207 family (including STxH238 and STxH239).

Prior to this fix, when the U-Boot command "reset" was executed, the ST40
CPU would force a "double-fault" condition, that reset the ST40 core.
However, this reset did not propagate to the rest of the system, and
hence the (off-core, on-chip) FLASH boot-controllers (for NAND and SPI)
were not reset correctly. As a result, the boot flash device was logically
"missing" after the reset, and consequently the re-boot failed!

This commit explicitly enables the ability for the ST40 resets (both
manual + watchdog) to be routed *OUT* of the ST40 core, and hence,
to propagate to the rest of the system, resulting in a re-bootable system.

Note, various users have seen different unexpected issues when trying
to reboot the system, it is believed that all of these different issues
should now all be fixed by this one commit. In particular, please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=19934
	https://bugzilla.stlinux.com/show_bug.cgi?id=20100

This has been tested on several different boards. Testing took place
with both boot-from-NAND, and boot-from-SPI. In addition, resets using
both the standard "double-fault" mechanism, and the ST40 watchdog were
both tested and found to work. No issues were noted.

Testing was undertaken using the following boards:
	B2057A board (STiH207-HDKAST N047) and
	B2064A board (STiH239-HDKA003 E009) and
	B2064B board (STiH239-HDKB200 E075) and
	B2067A board (STiH238-HDKA003 E013).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>