[ST40] Fix cases where DECLARE_GLOBAL_DATA_PTR was not declared as global
This commit is making the same set of changes, specifically for the ST40
architecture, as per commit 1218abf1b5817a39a82399b4b928b00750575bda.
That is, homogenizing DECLARE_GLOBAL_DATA_PTR to be of "file-scope",
and not of "function-scope".

In addition, a liberal sprinkling of the addition of the "const" qualifier
was also applied to the variable "bd" (of type bd_t), in the name of
"const-correctness", for the ST40 files at the same time.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>