[ST40] Replaced EMI clock with 3 clocks: UART, TMU, and SSC
The clocking infrastructure on the ST40 has now got very complicated,
and U-Boot sources (for ST40) were no longer able to cope with this
increase complexity, without some form of re-factoring exercise.

In the good old days, there was just a single "peripheral" clock, then a
"comms" clock was added.  At that point, there were just two distinct
clocks that we needed to deal with:
	The "comms" clock was used by: ASC, SSC, Watchdog(Comms)
	The "peripheral" clock was used by: TMU, SCIF, Watchdog(ST40)

Now, there may be two different "comms" clocks - one "system" clock,
and a second one for the SBC (Stand-By Controller) domain, and these two
"comms" clocks typically run at different frequencies. So, (for example)
the UART divider value to be used depended on which "comms" clock is
actually being used, depending if the ASC is in the SBC domain or not!

It is not required for U-Boot to have a very sophisticated appreciation
of the clock infrastructure, and their myriad of interdependencies and
their hierarchy. However, U-Boot does need to know how to obtain the
actual clock frequencies for each class of peripheral, and know how to
compute the divisors.  To this end, it has been decided to have a few
simple functions which will return the raw frequency for each peripheral
that needs to know its clock frequency.

Hence we now have the following new functions:

	stm_get_uart_clk_rate()	-> UART for the ASC (serial console)
	stm_get_tmu_clk_rate()	-> TMU  for the Timers (udelay())
	stm_get_ssc_clk_rate()	-> SSC  for the SPI via SSC

Each of these functions simply returns the appropriate frequency in the
"bd_info" structure (scaled to be in Hertz).

The key point, is that we have moved from being clock-centric (where one
clock serves many different IP blocks), to being block-centric (where
one clock serves only one IP block).  This is slightly wasteful if two IP
blocks use the same physical clock, but it gives greater flexibility if
that association changes in the future! Given the relatively few clocks
that U-Boot needs to know about, this seems to be the best option.

This has been built on all currently supported ST40 board, and executed
on a single board: B2057A board (STiH207-HDKAST N047). Testing comprised
just ensuring the serial console (using the UART clock), and the "sleep"
command (using the TMU clock) were both working correctly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>