[ST40] Only test 48-bits of the H/W MAC address for validity.
Fixed a minor issue with the detection of invalid H/W MAC addresses
in the "ROM" as seen by the MAC. The actual MAC address is 48 bits,
however the code compared 64-bits (2 * 32-bit registers). Hence the
code previously assumed that the "extra" 16 bits were all zero - which
is not necessarily true!

Now, the code will test only 48-bits, and it will determine the MAC
address as being invalid, only if these 48 bits are all ones.

This should remove the slightly annoying diagnostic, but almost ubiquitous
warning, that the MAC addresses do not "match":
	Warning: MAC addresses don't match:
	HW MAC address:  FF:FF:FF:FF:FF:FF
	"ethaddr" value: ...

Tested on one ST40 board (the B2000).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>