[ST40] Ensure HARD reset of PHY done after the MAC initialization
Following on from https://bugzilla.stlinux.com/show_bug.cgi?id=20554,
it was apparent that U-Boot for ST40 was inconsistent on the relative
order that the Network MAC and PHY were initialized. That is, for some
boards, the MAC was initialized before the PHY, and on other boards it
was the other way about. Demonstrably, this common inconsistency does
not normally seem to have been a noticeable problem, so it is slightly
surprising that after all these years, it is only now surfaced as a
real issue.  However, an issue it is!

The perceived wisdom is that the hard reset of the PHY (by toggling
the PHY_RESET line via a PIO under software control) should *always*
be performed *AFTER* the MAC initialization (the PIO routing and SYSCONF
initialization for the GMAC).  For completion, the soft-reset of the PHY
(using the SMI) should clearly always be performed after the hard reset
of the PHY.

The aforementioned bugzilla included a small patch (attachment #19574)
for the one board (HDK-7108) in question.  This commit simply modifies
the remaining ST40 board-specific files to ensure that the hard reset
of the PHY is performed after the MAC initialization. In addition, a
few comments have also been added/modified to make this paradigm shift
clearer in the source files. The soft PHY reset was always done after
the hard PHY reset anyway, so no changes there.

Thanks to Jon Frosdick for analysing and highlighting this issue.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>