[ST40] Rename Board from "HDK-H225" to "B2042".
It now recognised that all new (from circa 2012) boards from
STMicroelectronics, would be easier to identify uniquely using the new
"Bxxxx" naming convention (the first such board being the B2000).

The "HDK-H225" board is also referred to as the "B2042" (according to
the silk screen text).

Previously, the "HDK-H225" board (aka STiH225-HDK), was configured using
"hdkh225_config" or "hdkh225se_config". In hindsight, it has been decided
to "rename" this board to call it the "B2042". Hence the configuration
for it is now "b2042_config" or "b2042se_config". As a result, various
files have been correspondingly renamed, and a few files' contents have
also been updated to reflect the new naming convention for this board.

Testing by compiling only, as there is no substantive, or logical change.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>