[ST40] Standardize "help" messages for ST40-specific commands
This commit tidies-up, and standardizes the "help" commands, for the three
ST40-specific commands ("pmb", "copybbt", and "st_mtest") to make them
consistent with the other changes that were included in the "v2009.03"
release. For further details, please see:

	commit 62c3ae7c6ef215b1afa614abdf61acf077752207
	commit 2fb2604d5c20beb061b0a94282b7f6eb14d00cb8

Note: To test this, the following macros need to be defined:

	CONFIG_ST40_SE_MODE
	CONFIG_CMD_NAND
	CONFIG_SYS_NAND_SKIP_BAD_BLOCKS_ON_RELOCATING
	CONFIG_CMD_STTESTS

Testing comprised building U-Boot for the B2042 board, and running:

	U-Boot> help
	U-Boot> help nand ; help pmb ; help copybbt ; help st_mtest
	U-Boot> nand ; copybbt ; st_mtest

Ensuring that everything "looked" consistent.

All testing was performed on a single B2042 board "STiH225-HDK10ST E004".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>