[ST40] Add CONFIG_USB_EHCI to ST40 SoC-specific files in cpu/st40
With the recent addition of EHCI to the previous OHCI support for USB in
the current ST40 tree, we need to make the ST40 tree "EHCI" aware. This
commit ensures that the SoC-specific USB initialization functions (e.g.
functions of the form "<xxxx>_usb_init()") will now be conditionally
compiled in, if *either* OHCI or EHCI is enabled.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>