[USB] Add support for PHYSICAL addresses with EHCI driver
Prior to this commit, the EHCI driver was predicated on the assumption
that the (virtual) address as seen by the CPU, was the same (physical)
address as seen by the EHCI H/W controller. For a lot of systems this
equivalence holds, but not for the SuperH architecture, as used on ST40
SoCs from STMicroelectronics.

Due to the nature of the SoCs, the physical address of RAM is not the
same as the virtual address that the CPU uses, hence, one needs to "map"
some buffers from a virtual address space, into a physical address space.
Sometimes, one needs to use the virtual address, and sometimes one needs
to use the physical address for a given buffer, so the driver code needs
to manage this "duality" properly (as well as the caching implications).

Using the #define macro PHYSICAL_ADDR() in various places, the driver
will now use the correct address, when a H/W physical address needs to
be used.  Without this essential change, then the EHCI driver can not
be used on ST40 SoCs from STMicroelectronics.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>