[EHCI] do not call invalidate_dcache_range() with NULL pointers
For EHCI, do not try and invalidate the data-cache range, if the buffer
to be invalidated, is the same address as the NULL pointer!

This is not necessarily a real problem, but whilst writing the function
"invalidate_dcache_range()" for ST40, it was noticed that the function
was sometimes called with 'start' and 'stop' addresses of zero. It is
probably safer (and faster) not to do this...

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>