[ST40] Conditionally call OHCI cache manipulation functions
The USB 1.x OHCI driver is "immature" inasmuch as it does not have
any explicit code to purge/invalidate the data caches.  When OHCI
was initially ported for the ST40 SoCs, a pair of ST40-specific cache
functions ("sh_enable_data_caches()" and "sh_disable_data_caches()")
to enable/disable the data caches were introduced, to "turn-off"
the data caches, when executing "critical" code. This is sub-optimal,
(and a bit of a kludge) but it worked!

Now with the newer USB 2.0 EHCI driver updated to deal with the data
caches correctly, these functions are no longer needed to be called
for EHCI (but are still needed for OHCI). Unfortunately, two generic
USB files (common/cmd_usb.c, and common/usb_storage.c), were originally
modified to call the aforementioned cache functions unconditionally (in
the ST40 tree). So, for EHCI we do *not* want to call them, but for OHCI
(on ST40) we do *still* want to call them.

The present solution has been to surround all uses of these functions
with conditional complication, as follows:

	#if defined(CONFIG_ST40_STM_USB) && defined(CONFIG_USB_OHCI_NEW)
		sh_disable_data_caches();
		sh_enable_data_caches();
	#endif	/* CONFIG_ST40_STM_USB && CONFIG_USB_OHCI_NEW */

Ideally, these functions should be totally removed, and the OHCI driver
"taught" to manage the data caches directly (as the EHCI does). In the
mean time, this solution will need to suffice...

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>