[ST40] Add flush_dcache_range() and invalidate_dcache_range() for ST40
This commit adds the following two ST40-specific cache manipulation
functions to the STMicroelectronics' ST40-tree:

	flush_dcache_range(),
	invalidate_dcache_range().

These functions are used by the EHCI (USB 2.0) driver. Technically, both
these functions were already present in the code, but one was compiled
out, and the other was "static", with slightly different parameters. Both
these functions are now exported (as "extern"), and both take parameters,
which represent the "start", and the "stop" addresses of the range of
cache addresses to be manipulated.

In addition, a comment has been added to make it clear that the
"stop" address is really the byte *beyond* the last byte to be
flushed/invalidated. In other words, the address of the upper bound is
nominally excluded.

The previously exported function flush_cache() is still present, and now
calls flush_dcache_range(), after changing the domain of the parameters
from "start"+"size" to "start"+"stop".

Finally, the code has been simplified, in as much as there is no need to
force the "stop" address to be cache-line aligned. After some analysis,
it it obvious that if "start" is cache-line aligned, then "v" will always
be cache-line aligned, and we only need to keep looping while v < end.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>