[ST40] Complete adding EHCI capability for ST40 SoCs
Update all the ST40-specific board configuration files to be "EHCI"-aware.
That is, add a "#if 0 ... #else ... #endif" compilation comb to select
if we want to use the slower OHCI (USB 1.x), or the faster EHCI (USB 2.0)
controller when using the USB port on ST40 SoCs.

NOTE: In order to maintain full compatibility, prior to the addition of
support for the EHCI, then by default, all ST40 SoCs board configurations
will continue to use OHCI. That is, the board-specific config.h file
needs to be manually edited to use the new EHCI instead of OHCI ("#if 0"
changed to "if 1").

Typically, for OHCI, the following should all be defined:

	CONFIG_USB_OHCI_NEW
	CONFIG_SYS_USB_OHCI_CPU_INIT
	CONFIG_SYS_USB_OHCI_REGS_BASE
	CONFIG_SYS_USB_OHCI_SLOT_NAME
	CONFIG_SYS_USB_OHCI_MAX_ROOT_PORTS

Typically, for EHCI, the following should be defined instead:

	CONFIG_USB_EHCI
	CONFIG_USB_EHCI_STM

Tested on a single PDK7105 board (STi7105-SDK-V1.2-022), using a single
USB port (the default one), with *both* OHCI and EHCI controllers, in
*both* 29-bit mode, and 32-bit modes. The USB port was connected to a
4-port USB hub, with 3 downstream devices connected, two of which were
mass storage devices. No issues were noted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>