[ST40] Enable EHCI (USB 2.x) by default on PDK7105 board
For the PDK-7105 board, change the default USB controller to be the EHCI
(USB 2.0) one, instead of the OHCI (USB 1.x) one.

This is due to a sufficient amount of testing now having been undertaken
on one such board, to indicate that risk of migrating to EHCI is now
sufficiently low, to change the default mode, for (at least) this
specific board.

All testing was on a single PDK7105 board (STi7105-SDK-V1.2-022),
using *both* the USB ports (front and rear), with *both* OHCI and EHCI
controllers, in *both* 29-bit mode, and 32-bit modes. The USB port was
connected (in turn) to a 4-port USB 2.0 hub, with 3 downstream devices
connected, two of which were mass storage devices. No issues were noted.
Testing included using: "usb tree", "ext2ls", and "fatload".

WARNING: It should be noted that the EHCI controller (USB 2.0) does *not*
work directly with legacy USB 1.x only devices.  Hence, if an (old) USB
1.x device is to be used, then the configuration file should be modified
to select the OHCI controller instead. Most modern devices will work with
a EHCI controller, so this should be a rare issue, only for old devices.

NOTE: Although USB 1.x devices can not be used, when connected directly
to the EHCI controller, such devices are usable, if connected via a USB
2.0 HUB. Hence, this allows both USB 1.x and USB 2.0 devices to be seen by
the EHCI controller, as long as they are connected through a USB 2.0 hub.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>