spear3xx: Add pinmux support
The pinmux layer for spear family of devices exports three APIs for the board.
<soc/arch>_pins_default:
	This routine puts all the SoC pins in a default (safe) state. This API
	can be used by the respective boards to place all pins in a safe mode
	before going ahead with initializing with pinmux
<soc/arch>_select_mode:
	This routine selects a particular SoC mode. It is particularly for those
	SoCs which have several modes one of which can be selected. spear300 and
	spear320 are such SoCs.
<soc/arch>_enable_pins:
	This routine enables the desired pins. It accepts two arguments. First
	is 'ip' which indicates which controller pins are to be enabled and the
	second is 'mode' in which this ip has to run eg. ip can be PMX_ETH0 and
	mode can be PMX_ETH_RGMII/PMX_ETH_MII etc

Additionally, this patch also defines mdio_get_control for spear3xx which also
falls in the purview of pinmux. This is defined because spear310 and spear320
devices need to control mdio lines dynamically because these are shared. Define
the mdio_get_control routine to achieve this.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>