[ST40] Refactor PMB initialization code, to ease maintenance
This is essentially a janitorial change to move all the 32-bit mode PMB
initialization, of the first 4 PMB slots (PMB[3:0]), into one small
region in "start.S", and not have repetition of largely similar code
spread across each board-specific chunk in that file.

To aid in this task, the "pmb.h" file was also updated to define the
following two macros appropriately, depending on the amount of SDRAM
the system is configured to use (i.e. with CONFIG_SYS_SDRAM_SIZE):

	#define CONFIG_SYS_ST40_LMI_NEEDS_2_PMB_ENTRIES
	#define CONFIG_SYS_ST40_LMI_PMB_PAGE_SIZE

Only minimal testing, on a single HDK-7108 board (STi7108-HDK10ZA-N206).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>