[ST40] Make sh_toggle_pmb_cacheability() more extensible
The (assembler) function sh_toggle_pmb_cacheability(), made a fundamental
assumption that to convert from a cached PMB mapping, to an un-cached
PMB mapping, you simply OR-ed the PC with 0x10000000, and to perform
the reverse conversion, you correspondingly AND-ed with the inverse of
that mask. This is fine, when we know that the cached region starts at
0x80000000, and the un-cached region starts at 0x90000000.

With the anticipation of supporting the full 512 MiB of cached SDRAM,
the previous assumption is no longer viable, and a more robust, and more
flexible solution will soon be needed.

This commit essentially makes the code more robust, and extensible by
using an XOR operation instead to switch between cached and un-cached
regions in the PMB.  This will help in the future, when the mapping from
the cached/un-cached domains does not permit a simple OR/AND operation
with just *one* mask to facilitate the transitions of the PC. Whereas,
we can easily switch with a single XOR operation, and a single mask.

Only minimal testing, on a single HDK-7108 board (STi7108-HDK10ZA-N206).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>