[ST40] Support 512MiB of (cached) SDRAM on ST40 Boards
Previously, U-Boot on 32-bit mode on the ST40 mapped a maximum of 256MiB
of SDRAM twice in PMB virtual space - once cached (at 0x80000000),
and once uncached (at 0x90000000). In practice, this was not a real
problem on systems with 512MiB (or more) of SDRAM, as long as U-Boot,
a linux kernel, and the ramdisk could all fit into the first 256MiB of
RAM, which was almost always the case.

However, recently, there has been a need to support 512MiB of SDRAM
directly in U-Boot. Previously the entirety of CONFIG_SYS_SDRAM_SIZE of
SDRAM was mapped TWICE (once cached, and once uncached). Mapping 512MiB
twice via the PMB would use 100% of the 1.0 GiB PMB virtual address space,
leaving no free address space for any other (e.g. FLASH) memory-mapped
devices that need to use the EMI. Hence, mapping the entirety of the
512MiB of SDRAM twice, is simply not a viable option in U-Boot - an
alternative solution is required!

With this commit, if CONFIG_SYS_SDRAM_SIZE is set to 512MiB of SDRAM,
then instead, we create a single cached 512MiB mapping (at 0x80000000)
for the SDRAM, and we also map a single 16MiB of PMB virtual space on top
of the last 16MiB of mapped physical RAM. This 16MiB uncached mapping of
SDRAM is expected to be located in the PMB's virtual space at 0xBF000000.

Note: Linux typically also reserves the last PMB 16MiB virtual page
(at 0xBF000000) for itself during booting, and U-Boot must ensure that
this page is NOT valid when transitioning into linux. It is safe for
U-Boot to use this same page (as an un-cached mapping), as long are we
are certain to invalidate the mapping prior to the transition to linux.

i.e. For 512 MiB, we create 2 PMB mappings, which ultimately will be:

	Slot	Page-Size	Caching	  Virtual-Address
	----	---------	-------	  ---------------
	PMB[0]	512 MiB		CACHED	  0x80000000..0x9fffffff
	PMB[2]	 16 MiB		UN-cached 0xbf000000..0xbfffffff

The entirety of the 16 MiB page (PMB[2]), will *PHYSICALLY* be the same
as the last 16 MiB region in the 512MiB page. That is, the same 16 MiB of
physical memory will be mapped twice (once cached, and once un-cached)
in the PMB virtual area. This area that is mapped twice should include
U-Boot's ".text" section, i.e. the memory around "TEXT_BASE".

For further details, please see "include/asm-st40/pmb.h".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>