[ST40] Minor tidy-up, moving 2 macros from board-specific to CPU-specific
Move the definitions of the following two #define macros:

	#define CONFIG_SYS_SE_UNCACHED_BASE
	#define CONFIG_SYS_SE_SDRAM_WINDOW

From a board-specific configuration file, into the CPU-specific include
file "include/asm-st40/addrspace.h".

In SE (32-bit) mode, the values of these macros are generally fixed,
and based on the amount of SDRAM is usable in the PMB. This commit should
make maintenance on these easier.

The user may still define these values in the board-specific configuration
files, and they will (as one would expect) over-ride the default values
in the CPU-specific file.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>