[ST40] Move NAND FLASH to EMI_CSA# on STxH205 family of boards
On the set of boards which uses SoCs in the STxH205 family of devices from
STMicroelectronics, it is permissible, and advisable that the NAND device
is selected using "Chip Select 'A'" (i.e. EMI_CSA#). Perhaps surprisingly,
this association does not present any significant issues, even if SPI
serial FLASH is on EMI_BANK0 (which is traditionally associated with
EMI_CSA#), even when booting from SPI serial FLASH.

Hence, on these class of boards, we now *always* have NAND on EMI_CSA#,
irrespective if we are booting from NAND, or SPI serial FLASH. In fact,
later revisions of the aforementioned boards will even remove the "switch"
(e.g. SW4, or JF3), which allowed NAND_CS# to be tied to either EMI_CSA#
or EMI_CSB#. This new paradigm simplifies the software configuration
slightly, and should be deployed on all such board henceforth.

Tested (boot-from-NAND, and boot-from-SPI) on the following boards:

	B2057A board	(STiH207-HDKAST N001)
	B2064B board	(STiH207-HDKB-00 E155)
	B2067A board	(STiH238-HDKA003 E013)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>