SPEAr1340 evb padmux: Disable SSP chipselects cs2 and cs3
On SPEAr 1340 Rev-1.0 boards, ssp chipselect pins cs-2 and cs-3 are used
for DDR-CKE and DDR-RST, so enabling these pins causes U-boot
initialization to fail.

On board revisions Rev-2.1 and above the same pins are used for DDR as
well, but resistors R127 and R128 are not loaded, so no issues occur.

Disable these pins, so that U-Boot can be used for Rev-1.0 boards as
well.

Signed-off-by: Vinit Shenoy <vinit.shenoy@st.com>
Reviewed-by: Vipin Kumar <vipin.kumar@st.com>