spear/spl: Add generic getboottype() routine
This patch replaces multiple routines each named as xxx_boot_selected() with a
single getboottype() routine which returns a bitmask representing the device
from where it can boot.

This bootmask is independent of the platfrom on which the code is running. There
are 31 possibilities in a u32 returned type with 32nd bit representing
unsupported booting type.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>