[ST40] Add support for defining the poke-table with ABSOLUTE addresses
Previously, the "poke-table" was specified exclusively using PIC (position
independent code), as the U-Boot binary (u-boot.bin) could be located
almost *anywhere* in the CPU's memory map.

However, there is a recent requirement for the poke-table to be specified
using a fixed (absolute) address, irrespective of the actual location
of the u-boot.bin image! This new capability is now implemented using
the following 2 new CONFIG_SYS_xxx macros:

	#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_BEGIN
	#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_END

These specify the beginning of the poke-table, and the (byte beyond
the) end of the poke-table to be used.  If using this specific feature,
then please note the following points:

	1) Both addresses *must* be cached addresses (e.g. P1).  This code
	will not enforce this requirement!

	2) Both addresses *must* be 4-byte aligned.  This code will not
	enforce this requirement!

	3) The code will be executed only in 29-bit mode, so please
	ensure the addresses are chosen accordingly. Typically, the
	addresses will be in the P1 region, for a Physical address,
	which is in the EMI Boot Controller space (i.e. a cached address
	of the FLASH on EMI_BANK_0).

	4) If using this on a SoC with a "direct-mapped" cache, (instead
	of 2-way set associative), then additional care is required
	to avoid any cache lines for the poke-table being prematurely
	evicted due to "cache-thrashing"!

	5) As a minimum, the first 12-bytes (at the start address) will
	always be read, in poke_loop(), so the addresses specified must
	be valid, with no side-effects!

	6) In all cases, the referenced poke-table should terminate
	with a special end marker, to denote the end of the poke-table
	(this was an opcode of ZERO).  The macro "END_MARKER" can be
	used from the assembler.

	7) If an "empty" poke-table is required, then the empty
	poke-table referenced, should be at least a 12-byte region of
	all zeros. Please see the definition of the macro "END_MARKER"
	for a suitable definition (in "include/asm-st40/pokeloop.h").

This was successfully tested with no issues noted, on a single HDK-7108
board (STi7108-HDK22ST-E835) - once without the new macros defined
(i.e. with PIC); and once with the new macros defined as follows:
	#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_BEGIN   0x800005a0
	#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_END     0x80002838

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>