[ST40] rename _uboot_end_data to __uboot_bin_size, and make more robust.
Previously, a linker symbol ("_uboot_end_data") was used to denote the
"end of the .data section", and was subsequently used to calculate the
size of the raw binary "u-boot.bin" image, with:

	monitor_flash_len = (ulong) & _uboot_end_data - TEXT_BASE;

That calculation is fine, as long as the load memory address (LMA)
always agrees with the virtual memory address (VMA). However, if one
were to use the section attribute "AT()" to have distinct LMA and VMA
for a given section, then the wrong value would subsequently be computed.

This commit makes that calculation more robust, by defining instead:

	__uboot_bin_size = LOADADDR(.bss) - LOADADDR(.text);

In summary, this commit removes the linker symbol "_uboot_end_data",
and replace it with the linker symbol "__uboot_bin_size" instead. Also
it make the value of the new symbol more "robust", when LMA != VMA.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>