[ST40] Update u-boot.lds to allow data to be relocated to __USE_VMA_DATA
Both the "u-boot" and "u-boot.bin" images contain the following
allocatable and loadable sections:
	".text", ".rodata", ".data", and ".u_boot_cmd".
Whereas, the STACK, HEAP, and the ".bss" sections are all *not*
allocatable/loadable, hence they are not physically present in these
images.

When booting from GDB, by default, GDB loads all the loadable sections
into RAM, using their LMA (Load Memory Address). Normally, the LMA is
the same as the VMA (Virtual Memory Address).

When booting from FLASH, the "u-boot.bin" binary is burned into FLASH
in its entirety, and it copies itself as a homogeneous binary blob from
FLASH to RAM, and once relocated, runs directly out of RAM.

When booting from RAM, U-Boot is essentially a Secondary Boot Loader
(SBL), and some other "entity" has performed the duty of a primary boot
loader (PBL, or another booting mechanism). In this case, U-Boot does not
need to perform the pokes in the "poke-table", or even relocate itself.

However, boot-from-RAM, has one further (optional) complexity. It is
possible that when U-Boot has been coped to RAM, that the region of
RAM containing the u-boot.bin image is then marked as READ-ONLY by
the underlying hardware system, and U-Boot should not relocate itself,
but run in-place, directly out the RAM. This is not a problem for the
".text", and the ".rodata" sections, as (in general) these should be
read-only. It is not a problem for the STACK, HEAP, and the ".bss"
sections, as these should never be positioned in a read-only region
of RAM. However, this presents a significant problem for the ".data",
and the ".u_boot_cmd" sections, which must be stored in a READ-WRITE
region of RAM, but are also physically stored in the u-boot.bin binary
image which is now in a READ-ONLY region of RAM!

To solve this last problem, we need to "re-relocate" the READ-WRITE
sections that are allocated/loadable from RAM to RAM when booting from
RAM. To do this, we need to do a few things:

	1) place ".data" at a VMA where the CPU will normally access it

	2) place ".data" at a LMA where it physically abuts after ".rodata".

	3) when U-Boot starts, copy ".data" from the LMA to the VMA.

	Note: for ".data", read all sections that need to be read-write.
	Note: this commit takes care of points #1, and #2 (but not #3).

To control if the "normal" (VMA==LMA) placement scheme is used, or if a
"special" (VMA!=LMA) placement scheme is to be used, a new linker symbol
"__USE_VMA_DATA" will be used, with the following semantics:

	if (__USE_VMA_DATA == 0)
	{
		use the "normal" (or traditional) placement scheme.
		place ".data" VIRTUALLY+PHYSICALLY at end of ".rodata".
		i.e. VMA == LMA
	}
	else
	{
		use a "special" (non-normal) placement scheme.
		place ".data" PHYSICALLY at end of ".rodata" (LMA).
		place ".data" VIRTUALLY  at __USE_VMA_DATA (VMA).
		i.e. VMA != LMA
	}

The user will need to edit the "cpu/st40/u-boot.lds" file directly,
and define "__USE_VMA_DATA" as appropriate for their specific system.

The default is:

	__USE_VMA_DATA = 0;	/* use the NORMAL placement scheme */

Hence, this commit (as it stands) should *not* make any logical changes
to U-Boot's behaviour at all, expect that it adds a few more globally
visible linker symbols.

If "playing" with this capability, the following information may be
helpful. By default GDB uses the LMA addresses for loading images to
targets.  However, if a section has a different LMA and VMA addresses,
then unless steps are made to copy the sections in the LMA to be in the
VMA on the target, then the program will fail to execute correctly! You
can force GDB to use VMA addresses (instead of LMA), by using the
following GDB commands:
	(gdb) load ./u-boot vma
	(gdb) compare-sections vma
Note: You do need to specify the Elf file name (no default)!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>